Addressed Bin's Review Comments

+++ source/common/network/io_socket_handle_impl.h
@@ -82,7 +82,7 @@ private:
   // size when receiving a packet. It is possible for a received packet to contain both IPv4
   // and IPV6 addresses.
   const size_t cmsg_space_{CMSG_SPACE(sizeof(int)) + CMSG_SPACE(sizeof(struct in_pktinfo)) +
-                           CMSG_SPACE(sizeof(struct in6_pktinfo)) + CMSG_SPACE(sizeof(int))};
+                           CMSG_SPACE(sizeof(struct in6_pktinfo)) + CMSG_SPACE(sizeof(uint16_t))};
 };
 
 } // namespace Network
